==> notebooks/01.Fixing_Structure.ipynb <==
import pandas as pd
from pathlib import Path
from shutil import rmtree
from pprint import pprint

root = Path.cwd().parent
data = root / 'data'
train = data / 'train'
val = train / 'val'
tm = data / 'raw' / '2021' / 'train_mini'

df = pd.read_json(data / '2018' / 'categories.json')
df = df[df['kingdom'] == 'Fungi']
df

for p in (tv / 'Fungi').iterdir():
    classrow = df[df['id'] == int(p.name)]
    new = f"{classrow['kingdom'].values[0]}_{classrow['phylum'].values[0]}_{classrow['class'].values[0]}_{classrow['order'].values[0]}_{classrow['family'].values[0]}_{classrow['genus'].values[0]}_{classrow['name'].str.split(' ').str[1].values[0]}"
    if not (p.parent / new).is_dir():
        (p.parent / new).mkdir()
    for f in p.iterdir():
        f.rename(p.parent / new / f.name)    

for p in (tv / 'Fungi').iterdir():
    if not p.name.startswith("Fungi"):
        p.rmdir()

def filter_dirs(path):
    for p in path.iterdir():
        if p.name.startswith('Fungi'):
            print(p)
            continue
        else:
            rmtree(p)


def rename_dirs(path, target_dir):
    for p in path.iterdir():
        for f in p.iterdir():
            if not (target_dir / p.name).is_dir():
                print(target_dir / p.name)
                (target_dir / p.name).mkdir()
            for f in p.iterdir():
                f.rename(target_dir / p.name / f.name)
            p.rmdir()

from shutil import rmtree

def remove_dirs_no_fungi(path):
    for p in path.iterdir():
        if not p.name.startswith("Fungi"):
            rmtree(p)

def rename_remove_dirs(path):
    for p in path.iterdir():
        old = p.name
        new = '_'.join(str(old).split("_")[1:])
        new = p.parent / new
        if not new.is_dir():
            new.mkdir()
        for f in p.iterdir():
            f.rename(new / f.name)
        p.rmdir()

map(filter_dirs, [tv, test, t, train,])

map(rename_dirs, [(v, train), (tm, train)])
