==> notebooks/02.Joining_Locations.ipynb <==
import pandas as pd
import json
from pathlib import Path
from IPython.display import display

root = Path.cwd().parent
data = root / 'training' / 'data' / 'raw'
base2018 = data / '2018'
base2019 = data / '2019'
base2021 = data / '2021'
%cd ../
import training.src.data_processing.train_preprocessing as tp

df2018 = tp.parse_2018_data(base2018)
df2021 = tp.parse_2021_data(base2021)
df2019 = tp.parse_2019_data(base2019)

df2019.columns

df2018.drop(['license', 'rights_holder', 'location_uncertainty', 'date', 'valid', 'user_id', 'latitude', 'longitude', 'supercategory'], axis=1, inplace=True)
df2019 = df2019.drop(['license', 'rights_holder'], axis=1).rename({'supercategory': "kingdom"}, axis=1)
df2021.drop(['license', 'rights_holder', 'image_id', 'location_uncertainty', 'date', 'latitude', 'longitude', 'supercategory', 'image_dir_name'], axis=1, inplace=True)

set(df2019.columns.tolist()).intersection(set(df2018.columns.tolist())).intersection(set(df2021.columns.tolist()))

df = pd.concat([df2018, df2019, df2021], ignore_index=True)

df['class_id'] = df.groupby('name').ngroup()
df = df.drop_duplicates(subset='file_name')

df[['dataset', 'dset', 'kingdom', 'name', 'family', 'phylum', 'order', 'genus', 'specific_epithet']] = df[['dataset', 'dset', 'kingdom', 'name', 'family', 'phylum', 'order', 'genus', 'specific_epithet']].astype('category')
df.to_csv(data / 'train.csv', index=False)

def parse_json(filepath, istest=False, categories=None):
    with open(filepath, 'r') as f:
        res = json.load(f)
    info = pd.DataFrame.from_dict(res['info'], orient='index')
    images = pd.DataFrame(res['images']).set_index('id')
    if not istest:
        annotations = pd.DataFrame(res['annotations']).set_index('id')
        if categories:
            categories = pd.DataFrame(res['categories']).set_index('id')
            return info, images, annotations, categories
        return info, images, annotations

    return info, images

def join_dataframes(images, annotations, categories, locations=None):
    categories = categories[categories['supercategory'] == 'Fungi']
    categories = categories.rename(columns={'id': 'category_id'})
    if locations is None:
        df = categories.merge(annotations, right_on='category_id', left_index=True, how='inner').merge(images, left_on='image_id', right_index=True, how='inner')
    else:
        df = categories.merge(annotations, right_on='category_id', left_index=True, how='inner').merge(images, left_on='image_id', right_index=True, how='inner').merge(locations, left_on='image_id', right_index=True, how='inner')
    try:
        df = df.drop(['supercategory', 'kingdom', 'image_id', 'valid', 'license', 'rights_holder', 'user_id'], axis=1)
    except KeyError:
        df = df.drop(['supercategory', 'kingdom', 'image_id', 'license', 'rights_holder'], axis=1)
    finally:
        return df


def display_data(*args):
    for df in args:
        display(df)

## Parse JSON Data

tinfo2018, timages2018, tanno2018 = parse_json(data / '2018' / 'train2018.json')
vinfo2018, vimages2018, vanno2018 = parse_json(data / '2018' / 'val2018.json')
# testinfo2018, testimages2018 = parse_json(data / '2018' / 'test2018.json', istest=True)
tloc = pd.read_json(data / '2018' / 'inat2018_locations' / 'train2018_locations.json').set_index('id')
vloc = pd.read_json(data / '2018' / 'inat2018_locations' / 'val2018_locations.json').set_index('id')

with open(data / '2018' / 'categories.json', 'r') as f:
    res = json.load(f)
    cats = pd.DataFrame(res)

# tinfo2019, timages2019, tanno2019, tcat2019 = parse_json(data / '2019' / 'train2019.json', categories=True)

tinfo2021, timages2021, tanno2021, tcat2021 = parse_json(data / '2021' / 'train.json', categories=True)
vinfo2021, vimages2021, vanno2021, vcat2021 = parse_json(data / '2021' / 'val.json', categories=True)
# testinfo2021, testimages2021 = parse_json(data / '2021' / 'public_test.json', istest=True)

# 2018 Dataset

display_data(tloc, timages2018, tanno2018, cats)

val = join_dataframes(vimages2018, vanno2018, cats, vloc)
val['set'] = "validation"

train = join_dataframes(timages2018, tanno2018, cats, tloc)
train['set'] = 'train'
df = pd.concat([train, val]).reset_index(drop=True)

df['name'] = df['name'].apply(lambda x: '_'.join(x.split(' ')))
df['new_dirs'] = df.apply(lambda x: f"Fungi_{x['phylum']}_{x['class']}_{x['order']}_{x['family']}_{x['name']}", axis=1)
df['new_paths'] = df['new_dirs'] + '/' + df['file_name'].str.split('/').str[-1]

# target = data / '2018' / 'train_val'
# target.mkdir(exist_ok=True)
# for group in df.groupby('new_dirs'):
#     target_dir = target / group[0]
#     target_dir.mkdir(exist_ok=True)
#     for file in group[1]['file_name']:
#         file = data / '2018' / Path(file)
#         parent_dir = file.parent
#         file.rename(target_dir / group[1]['new_paths'])
#     parent_dir.rmdir()

df.loc[:, 'file_name'] = df['file_name'].str.split('/').str[-1]
df = df.drop(['category_id', 'new_paths'], axis=1)
df['specific_epithet'] = df['name'].str.split(' ').str[-1]
df['image_dir_name'] = "Fungi_" + df['phylum'] + '_' + df['class'] + '_' + df['order'] + '_' + df['family'] + '_' + df['genus'] + '_' + df['specific_epithet']
df = df.rename(columns={'lon': 'longitude', 'lat': 'latitude', 'loc_uncert': 'location_uncertainty'})
# df.to_csv(data / '2018' / 'train_val.csv', index=False)

# This is just fixing a screwed up made when joining the 2018 and 2021 datasets..
all_data = Path.cwd().parent / 'data' / 'train'
for f in all_data.iterdir():
    if f.is_file():
        genus = df[df['file_name'] == f.name]['new_dirs'].values
        if len(genus) > 0:
            genus = genus[0]
            if not (all_data / genus).exists():
                (all_data / genus).mkdir(exist_ok=True)
            f.rename(all_data / genus / f.name)
        else:
            print(f.name)

## 2021 Dataset

display_data(timages2021, tanno2021, tcat2021)

train2 = join_dataframes(timages2021, tanno2021, tcat2021)
val2 = join_dataframes(vimages2021, vanno2021, vcat2021)
train2['set'] = 'train'
val2['set'] = 'validation'
df2 = pd.concat([train2, val2]).reset_index(drop=True)

tpath = data / '2021' / 'train'
for p in tpath.iterdir():
    new_name = '_'.join(p.name.split('_')[1:])
    new_dir = tpath / new_name
    new_dir.mkdir()
    for f in p.iterdir():
        f.rename(new_dir / f.name)
    p.rmdir()

df2['image_dir_name'] = df2['image_dir_name'].apply(lambda x: '_'.join(x.split('_')[1:]))
df2 = df2.drop('category_id', axis=1)
df2['dataset'] = '2021'
df2['file_name'] = df2['file_name'].str.split('/').str[-1] 

df = pd.concat([df, df2], ignore_index=True)
df['file_path'] = 'Mushroom-Classifier/data/train/' + df['image_dir_name'] + '/' + df['file_name']
df['gcs_path'] = 'gs://mush-img-repo/train/ ' + df['image_dir_name'] + '/' + df['file_name']
df['class_id'] = df.groupby('name').ngroup()
df['date'] = pd.to_datetime(df['date'], format='mixed', utc=True)
df.to_csv(root / 'data' / 'train_val.csv', index=False)
