==> notebooks/06.Training.ipynb <==
import pandas as pd
import numpy as np
import tensorflow as tf
import matplotlib.pyplot as plt
import seaborn as sns
from pathlib import Path
import sys
# sys.path.append('..')

from config import CFG
from src.training import dataset, train, utils

root = Path.cwd().parent
data = root / 'data'

df = pd.read_csv(data / 'train.csv')
CFG.NUM_CLASSES = len(df['class_id'].unique())

strategy, tpu = train.tpu_test(CFG)
train.train(CFG, strategy)
