==> notebooks/08.MO-EDA.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path

root = Path.cwd().parent
data = root / 'data' / "MO"

names = pd.read_csv(data / 'names.csv', sep='\t', index_col=0, usecols=[0, 1, 5, 6])
images = pd.read_csv(data / 'images.csv', sep='\t', index_col=0, usecols=[0, 2, 5])
image_o = pd.read_csv(data / 'images_observations.csv', sep='\t', index_col=0)
loc_d = pd.read_csv(data / 'locations.csv', sep='\t', index_col=0,)
name_c = pd.read_csv(data / 'name_classifications.csv', sep='\t', index_col=0)

observations = pd.read_csv(data / 'observations.csv', sep='\t', index_col=0, parse_dates=[2])

observations.dropna(subset=['lat', 'long', 'alt'])

loc_d
